moves file, and functions import
